no need for looping through invoices for admin panel
this is different for payments.

note: there are a aside case for payment not covered here

if we were to have 2 lnd, and the payment would have been sent through the first lnd, and failed. then send again with the second lnd and succeed.

we may only show the failed payment where in fact the payment would have succeed.

I think we should rely on medici also for the payments