refactor: remove legacy 'usd'-related ledger transaction properties (#2191)
* chore: remove 'usd' props from ledger transactions type

* test: remove 'usd' props from ledger transactions type

* feat: convert to 'usd' amount in tx-history

This calculation would have previously been captured in the 'usd' and
'feeUsd' fields on transactions and then read out here. Since we now
only store cents on transactions, this calculation must explicitly be
done here.

This is likely a temporary change as the ideal state should be for the
price information to be captured directly in txn metadata instead of
being derived as an imprecise ratio of amounts persisted.

* chore: remove 'usd' props from ledger metadata

* refactor: rename 'feeUsd' variables to 'centsAmount'

* refactor: rename 'feeBtc' param to 'satsFee'

* refactor: change variable names to major/minor unit

* refactor: re-add usd/feeUsd/fee props and handle for admin txns