test: continue cleaning up receive ln integration (#3257)
* test: remove memo-spam test

This is a tx-history concern.

* test: add 1st trigger-interrupt ln-receive test

* test: fix transaction check to accommodate multiple passing transactions

* test: add 2nd trigger-interrupt ln-receive test

* temp: debug last trigger-interrupt test

We have an issue where the received invoice is being recorded 3 times
incorrectly. This is to show the issue in '.e2e-trigger.log' so that we
can debug and fix.

* fix: check re-fetched walletinvoice inside lock

When updating pending invoices we refetch the wallet invoice to make
sure it hasn't been paid as yet. This fixes a bug where we were instead
checking a previously fetched wallet invoice from outside the lock.

* revert: debug last trigger-interrupt test

This reverts commit d7d23968308dde3218904f2ed7682ccb69ab7697.